requirements.txt
# Scheduling library
schedule==1.2.0

# Data manipulation
numpy==1.23.5
pandas==1.5.3
PyYAML==6.0

# faker
faker==18.9.0

# duckdb
duckdb==1.1.3

pyarrow==12.0.1